chore: remove spaces

+++ src/main/scala/click/seichi/regenerateworld/domain/GenerationSchedule.scala
@@ -4,9 +4,9 @@ import java.time.ZonedDateTime
 import java.util.UUID
 import scala.util.Try
 
-// TODO: rename interval to minutesInterval
-// TODO: rename nextDate to nextDateTime
-case class GenerationSchedule(id: UUID, nextDate: ZonedDateTime, interval: Int, seedPattern: SeedPattern, worlds: Set[String]) {
+//TODO: rename interval to minutesInterval
+//TODO: rename nextDate to nextDateTime
+case class GenerationSchedule(id: UUID, nextDate: ZonedDateTime, interval: Long, seedPattern: SeedPattern, worlds: Set[String]) {
   // region Init
 
   if (interval <= 0) throw new IllegalArgumentException("The value of Interval must be more than 0")
